tests/test_config_loader.py
"""Tests for config_loader module."""

import os
import pytest
from pathlib import Path
import sys

# Add src to path
sys.path.insert(0, str(Path(__file__).parent.parent / "src"))

from config_loader import ConfigLoader


def test_config_loader_basic():
    """Test basic configuration loading."""
    # This is a placeholder test
    # Real tests would require a test config file
    pass


def test_env_var_resolution():
    """Test environment variable resolution."""
    # This is a placeholder test
    pass


def test_get_with_dot_notation():
    """Test getting config values with dot notation."""
    # This is a placeholder test
    pass
