requirements.txt
python-telegram-bot==20.7
prometheus-api-client==0.5.5
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.1
pyyaml==6.0.1
loguru==0.7.2
